// src/app/model/intransit.ts
export class Intransit {
  id: string='';
  productId: string='';
  productCode: string = '';
  toLocation: string = '';
  fromLocation: string = '';
  productWeight: string='';
  challanNumber: string='';
  driverContact: string = '';
  dateOfLoading: string = '';
  dateCreated: string = '';
  description: string = '';
  }